force dtype

+++ benchmark/run/numeric_benchmark.py
@@ -5,10 +5,10 @@ from benchmark.run.base import pretty_print, run_backends
 
 
 def generate_obj(size):
-    x = np.linspace(0, 1, size)
-    y = np.linspace(-1, 1, size)
-    z = np.linspace(0, 1, size)
-    values = np.random.rand(size)
+    x = np.linspace(0, 1, size, dtype=np.float64)
+    y = np.linspace(-1, 1, size, dtype=np.float64)
+    z = np.linspace(0, 1, size, dtype=np.float64)
+    values = np.random.rand(size).astype(dtype=np.float64)
     info = Information(
         "Numeric benchmark", "Data for numeric benchmark", 0, "fandreuz@cern.ch"
     )
